+++ apps/stripe/src/app/api/saleor/transaction-cancelation-requested/use-case-response.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from "vitest";
 
 import { getMockedSaleorMoney } from "@/__tests__/mocks/constants";
 import { mockedStripePaymentIntentId } from "@/__tests__/mocks/mocked-stripe-payment-intent-id";
-import { StripeAPIError } from "@/modules/stripe/stripe-payment-intent-api-error";
+import { StripeAPIError } from "@/modules/stripe/stripe-api-errors";
 import {
   CancelFailureResult,
   CancelSuccessResult,
@@ -58,7 +58,7 @@ describe("TransactionCancelationRequestedUseCaseResponses", () => {
           ],
           "amount": 0,
           "externalUrl": "https://dashboard.stripe.com/payments/pi_TEST_TEST_TEST",
-          "message": "Payment intent error - there is a problem with the request to Stripe API",
+          "message": "There is a problem with the request to Stripe API",
           "pspReference": "pi_TEST_TEST_TEST",
           "result": "CANCEL_FAILURE",
         }
