+++ apps/stripe/src/app/api/saleor/transaction-process-session/use-case-response.ts
@@ -7,9 +7,9 @@ import { SuccessWebhookResponse } from "@/modules/saleor/saleor-webhook-response
 import { generateStripeDashboardUrl } from "@/modules/stripe/generate-stripe-dashboard-url";
 import {
   StripeApiErrorPublicCode,
+  StripeApiErrors,
   StripeCardErrorPublicCode,
-  StripeGetPaymentIntentAPIError,
-} from "@/modules/stripe/stripe-payment-intent-api-error";
+} from "@/modules/stripe/stripe-api-errors";
 import {
   AuthorizationActionRequiredResult,
   ChargeActionRequiredResult,
@@ -67,7 +67,7 @@ class Success extends SuccessWebhookResponse {
 
 class Failure extends SuccessWebhookResponse {
   readonly transactionResult: ChargeFailureResult | AuthorizationFailureResult;
-  readonly error: StripeGetPaymentIntentAPIError;
+  readonly error: StripeApiErrors;
   readonly saleorEventAmount: number;
 
   private static ResponseDataSchema = createFailureWebhookResponseDataSchema(
@@ -81,7 +81,7 @@ class Failure extends SuccessWebhookResponse {
 
   constructor(args: {
     transactionResult: ChargeFailureResult | AuthorizationFailureResult;
-    error: StripeGetPaymentIntentAPIError;
+    error: StripeApiErrors;
     saleorEventAmount: number;
   }) {
     super();
