+++ apps/stripe/src/app/api/saleor/transaction-process-session/use-case.test.ts
@@ -17,7 +17,7 @@ import {
   BrokenAppResponse,
   MalformedRequestResponse,
 } from "@/modules/saleor/saleor-webhook-responses";
-import { StripeAPIError } from "@/modules/stripe/stripe-payment-intent-api-error";
+import { StripeAPIError } from "@/modules/stripe/stripe-api-errors";
 import { IStripePaymentIntentsApiFactory } from "@/modules/stripe/types";
 
 import { TransactionProcessSessionUseCase } from "./use-case";
