+++ apps/stripe/src/modules/stripe/stripe-api-errors.ts
@@ -6,7 +6,7 @@ import { BaseError } from "@/lib/errors";
 export const StripeCardErrorPublicCode = "StripeCardError" as const;
 export const StripeApiErrorPublicCode = "StripeApiError" as const;
 
-type StripePaymentIntentAPIError = InstanceType<
+export type StripeApiErrors = InstanceType<
   | typeof StripeCardError
   | typeof StripeInvalidRequestError
   | typeof StripeRateLimitError
@@ -18,23 +18,12 @@ type StripePaymentIntentAPIError = InstanceType<
   | typeof StripeUnknownAPIError
 >;
 
-export type StripeCreatePaymentIntentAPIError = StripePaymentIntentAPIError;
-export type StripeGetPaymentIntentAPIError = Omit<StripePaymentIntentAPIError, "StripeCardError">;
-export type StripeCapturePaymentIntentAPIError = Omit<
-  StripePaymentIntentAPIError,
-  "StripeCardError"
->;
-export type StripeCancelPaymentIntentAPIError = Omit<
-  StripePaymentIntentAPIError,
-  "StripeCardError"
->;
-
 export const StripeCardError = BaseError.subclass("StripeCardError", {
   props: {
     _internalName: "StripePaymentIntentsApi.StripeCardError" as const,
     publicCode: StripeCardErrorPublicCode,
     publicMessage: "", // filed by Stripe - they recommend to show it to the storefront user
-    merchantMessage: "Payment intent error - there is problem with processing of the card",
+    merchantMessage: "There is problem with processing of the card",
   },
 });
 
@@ -43,7 +32,7 @@ export const StripeInvalidRequestError = BaseError.subclass("StripeInvalidReques
     _internalName: "StripePaymentIntentsApi.StripeInvalidRequestError" as const,
     publicCode: StripeApiErrorPublicCode,
     publicMessage: "There is a problem with the request to Stripe API",
-    merchantMessage: "Payment intent error - there is a problem with the request to Stripe API",
+    merchantMessage: "There is a problem with the request to Stripe API",
     stripeCode: "",
     stripeParam: "",
   },
@@ -54,7 +43,7 @@ export const StripeRateLimitError = BaseError.subclass("StripeRateLimitError", {
     _internalName: "StripePaymentIntentsApi.StripeRateLimitError" as const,
     publicCode: StripeApiErrorPublicCode,
     publicMessage: "There is a problem with the request to Stripe API",
-    merchantMessage: "Payment intent error - there is a problem with the request to Stripe API",
+    merchantMessage: "There is a problem with the request to Stripe API",
   },
 });
 
@@ -63,7 +52,7 @@ export const StripeConnectionError = BaseError.subclass("StripeConnectionError",
     _internalName: "StripePaymentIntentsApi.StripeConnectionError" as const,
     publicCode: StripeApiErrorPublicCode,
     publicMessage: "There is a problem with the request to Stripe API",
-    merchantMessage: "Payment intent error - there is a problem with the request to Stripe API",
+    merchantMessage: "There is a problem with the request to Stripe API",
   },
 });
 
@@ -72,7 +61,7 @@ export const StripeAPIError = BaseError.subclass("StripeAPIError", {
     _internalName: "StripePaymentIntentsApi.StripeAPIError" as const,
     publicCode: StripeApiErrorPublicCode,
     publicMessage: "There is a problem with the request to Stripe API",
-    merchantMessage: "Payment intent error - there is a problem with the request to Stripe API",
+    merchantMessage: "There is a problem with the request to Stripe API",
   },
 });
 
@@ -81,7 +70,7 @@ export const StripeAuthenticationError = BaseError.subclass("StripeAuthenticatio
     _internalName: "StripePaymentIntentsApi.StripeAuthenticationError" as const,
     publicCode: StripeApiErrorPublicCode,
     publicMessage: "There is a problem with the request to Stripe API",
-    merchantMessage: "Payment intent error - there is a problem with the request to Stripe API",
+    merchantMessage: "There is a problem with the request to Stripe API",
   },
 });
 
@@ -90,7 +79,7 @@ export const StripePermissionError = BaseError.subclass("StripePermissionError",
     _internalName: "StripePaymentIntentsApi.StripePermissionError" as const,
     publicCode: StripeApiErrorPublicCode,
     publicMessage: "There is a problem with the request to Stripe API",
-    merchantMessage: "Payment intent error - there is a problem with the request to Stripe API",
+    merchantMessage: "There is a problem with the request to Stripe API",
   },
 });
 
@@ -99,7 +88,7 @@ export const StripeIdempotencyError = BaseError.subclass("StripeIdempotencyError
     _internalName: "StripePaymentIntentsApi.StripeIdempotencyError" as const,
     publicCode: "StripeApiError" as const,
     publicMessage: "There is a problem with the request to Stripe API",
-    merchantMessage: "Payment intent error - there is a problem with the request to Stripe API",
+    merchantMessage: "There is a problem with the request to Stripe API",
   },
 });
 
@@ -108,13 +97,11 @@ export const StripeUnknownAPIError = BaseError.subclass("StripeUnknownAPIError",
     _internalName: "StripePaymentIntentsApi.StripeUnknownAPIError" as const,
     publicCode: "StripeApiError" as const,
     publicMessage: "There is a problem with the request to Stripe API",
-    merchantMessage: "Payment intent error - there is a problem with the request to Stripe API",
+    merchantMessage: "There is a problem with the request to Stripe API",
   },
 });
 
-export const mapStripeCreatePaymentIntentErrorToApiError = (
-  error: unknown,
-): StripeCreatePaymentIntentAPIError => {
+export const mapStripeErrorToApiError = (error: unknown): StripeApiErrors => {
   switch (true) {
     case error instanceof Stripe.errors.StripeCardError:
       return new StripeCardError("Card payment error", {
@@ -149,43 +136,3 @@ export const mapStripeCreatePaymentIntentErrorToApiError = (
       });
   }
 };
-
-export const mapStripeGetPaymentIntentErrorToApiError = (
-  error: unknown,
-): StripeGetPaymentIntentAPIError => {
-  switch (true) {
-    case error instanceof Stripe.errors.StripeInvalidRequestError:
-      return new StripeInvalidRequestError("Invalid parameters provided to Stripe API", {
-        props: {
-          stripeCode: error.code ?? "",
-          stripeParam: error.param ?? "",
-        },
-      });
-    case error instanceof Stripe.errors.StripeRateLimitError:
-      return new StripeRateLimitError("Too many requests made to the API too quickly");
-    case error instanceof Stripe.errors.StripeConnectionError:
-      return new StripeConnectionError("There was a network problem between app and Stripe");
-    case error instanceof Stripe.errors.StripeAPIError:
-      return new StripeAPIError("Something went wrong on Stripe end");
-    case error instanceof Stripe.errors.StripeAuthenticationError:
-      return new StripeAuthenticationError("App can’t authenticate with Stripe");
-    case error instanceof Stripe.errors.StripePermissionError:
-      return new StripePermissionError("API key doesn’t have permission to perform this action");
-    case error instanceof Stripe.errors.StripeIdempotencyError:
-      return new StripeIdempotencyError("Idempotency key error");
-    default:
-      captureException(error);
-
-      return new StripeUnknownAPIError("Unknown error", {
-        cause: error,
-      });
-  }
-};
-
-export const mapStripeCapturePaymentIntentErrorToApiError = (
-  error: unknown,
-): StripeCapturePaymentIntentAPIError => mapStripeGetPaymentIntentErrorToApiError(error);
-
-export const mapStripeCancelPaymentIntentErrorToApiError = (
-  error: unknown,
-): StripeCancelPaymentIntentAPIError => mapStripeGetPaymentIntentErrorToApiError(error);
